pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift"
version = "0.1.0"
description = "Density-split Renormalized Inference and Field Theory"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "cosmoprimo",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["drift*"]
